enh(broker): New output unified sql (#112)
* New output unified_sql
* packaging is now made globally, collect is compiled and then each part is packaged.
* centreon-tests has been completed with tests specific to unified_sql.

REFS: MON-11220

+++ centreon-broker/storage/inc/com/centreon/broker/storage/stream.hh
@@ -51,7 +51,7 @@ class stream : public io::stream {
   struct metric_info {
     bool locked;
     uint32_t metric_id;
-    uint32_t type;
+    uint16_t type;
     double value;
     std::string unit_name;
     double warn;
